create DefaultExamplesFilter with the common filters from examples/filters and translator/filters
